+++ lib/src/screenshot_comparator.dart
@@ -1,2 +1,0 @@
-export '_screenshot_comparator_io.dart'
-    if (dart.library.js_interop) '_screenshot_comparator_web.dart';